refactor(core): promote DocumentId to dbflux_core::document_id (#102)
Move DocumentId struct (with new() and Default impl) from dbflux_ui into
a new dbflux_core::document_id module. Re-export from dbflux_core::lib.rs
and keep a pub-use shim at the original dbflux_ui path so all call-sites
remain source-compatible. No Cargo.toml changes (uuid already present).
