+++ crates/dbflux_ui/src/ui/document/types.rs
@@ -2,21 +2,7 @@
 
 use uuid::Uuid;
 
-/// Unique identifier for a document.
-#[derive(Clone, Copy, PartialEq, Eq, Hash, Debug)]
-pub struct DocumentId(pub Uuid);
-
-impl DocumentId {
-    pub fn new() -> Self {
-        Self(Uuid::new_v4())
-    }
-}
-
-impl Default for DocumentId {
-    fn default() -> Self {
-        Self::new()
-    }
-}
+pub use dbflux_core::document_id::DocumentId;
 
 /// Supported document types.
 #[derive(Clone, Copy, PartialEq, Eq, Debug)]
